Move queries around, use named queries
